FrontPower version 2.1
Back to charging thruster battery from house-bank.

Victron will have a very low threshold to ensure there is enough
time to charge thruster battery - even if house bank is low.
(In the past, the problem was that a low house bank caused input voltage
at Victron to never reach the threshold)

Drawback of always keeping Vicron 'alive' is that it will always drain
on house bank. Hope is that the drain will be negligble when thruster has
fully charged.